Several enhancements
* Fix type hinting for EmmyLua
* Rename stasis arc to macaroni arc
+ Recategorize macros
+ Arc position reversing macro
* Migrate most macros to use JayCurry

+++ editing/cut-arc-on-timing.lua
@@ -3,6 +3,8 @@ do
     local Class = require("rech.Class")
     ---@type rech.lib.Request
     local request = require("rech.lib.request")
+    ---@type rech.jaycurry.JayCurry
+    local JayCurry = require("rech.jaycurry.JayCurry")
     
     ---@class rech.editing.CutArcOnTiming
     local this = Class()
@@ -18,16 +20,37 @@ do
     end
 
     function this.activate()
-        local result = request.Event(EventSelectionConstraint.arc(), "Select arc to cut")
-        local arc = result.arc[1]
+        local arcs = JayCurry.query("arc:sel")
+        if #arcs.events.arc == 0 then
+            notifyWarn("Select at least one arc or traces!")
+            return
+        end
         local cutTiming = request.Timing(true, "Select timing")
-        local cutpos = arc.positionAt(cutTiming, false)
-        local commands = Command.create("Cut arc on timing")
-        commands.add(arc.delete())
-        local t1, t2 = arc.timing, cutTiming
-        commands.add(Event.arc(t1, arc.startXY, t2, cutpos, arc.isTrace, arc.color, arc.type, arc.timingGroup, arc.sfx).save())
-        t1, t2 = cutTiming, arc.endTiming
-        commands.add(Event.arc(t1, cutpos, t2, arc.endXY, arc.isTrace, arc.color, arc.type, arc.timingGroup, arc.sfx).save())
+        local commands = Command.create(string.format("%s (%s)", __MACRO_DISPLAY_NAME, __MACRO_ID))
+        for _,arc in ipairs(arcs.events.arc) do
+            commands.add(arc.delete())
+            local cutpos = arc.positionAt(cutTiming, false)
+
+            local t1, t2 = arc.timing, cutTiming
+            local arc1 = Event.arc(t1, arc.startXY, t2, cutpos, arc.isTrace, arc.color, arc.type, arc.timingGroup, arc.sfx)
+            commands.add(arc1.save())
+            t1, t2 = cutTiming, arc.endTiming
+            local arc2 = Event.arc(t1, cutpos, t2, arc.endXY, arc.isTrace, arc.color, arc.type, arc.timingGroup, arc.sfx)
+            commands.add(arc2.save())
+
+            Event.setSelection({arc})
+            local arctaps = JayCurry.query("arc:sel:arctap")
+            for _,arctap in ipairs(arctaps.events.arctap) do
+                commands.add(arctap.delete())
+                arctap = arctap.copy()
+                if arctap.timing <= cutTiming then
+                    arctap.arc = arc1
+                else
+                    arctap.arc = arc2
+                end
+                commands.add(arctap.save())
+            end
+        end
         commands.commit()
     end
